Move socket_functions impl
